franquicia-plus-be-master/requirements.txt
asgiref==3.7.2
boto3==1.26.39
botocore==1.29.39
django==4.2.7
django-colorfield==0.7.2
django-cors-headers==4.3.1
django-environ==0.8.1
django-filter==21.1
django-material-admin==1.8.6
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.0.0
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.13
sendgrid==6.11.0